+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### New Features
 - Add in exo-open to options when launching server and show dialog when cant be launched [#391]
 - Show progress dialog when using the Install Fabric API button when adding mods to an instance
+- Show progress dialog when adding a mod from modrinth to an instance
 
 ### Fixes
 - Remove remote error reporting when pack image couldn't be loaded